backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai-whisper==20231117
requests==2.31.0
cloudinary==1.36.0
google-api-python-client==2.108.0
pydantic==2.5.0
pydantic-settings==2.1.0
groq==0.32.0
bcrypt==4.1.2
